BIG COMMIT
I forgot commits are a thing! I've been working on the grid, the size seems to be an issue (I kinda forgot how Unity works with screen sizes; we definately need to scale up our pixel art...).
I have a grid of spot objects, now I need to create objects to sit on thos spots, and I need to get the controls working.